# api_Getway/docs/swagger.yaml
basePath: /
definitions:
...
  borrowers.Borrower:
    properties:
      book_id:
        type: string
      borrow_date:
        type: string
      created_at:
        type: string
      id:
        type: string
      return_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  borrowers.CreateBorrowerRequest:
    properties:
      book_id:
        type: string
      borrow_date:
        type: string
      return_date:
        type: string
      user_id:
        type: string
    type: object
  borrowers.CreateBorrowerResponse:
    properties:
      borrower:
        $ref: '#/definitions/borrowers.Borrower'
    type: object
  borrowers.GetAllBorrowersResponse:
    properties:
      borrowers:
        items:
          $ref: '#/definitions/borrowers.Borrower'
        type: array
    type: object
  borrowers.GetBorrowerByIdResponse:
    properties:
      borrower:
        $ref: '#/definitions/borrowers.Borrower'
    type: object
  borrowers.GetBorrowingHistoryResponse:
    properties:
      currently_borrowed:
        description: Books currently borrowed by the user
        items:
          $ref: '#/definitions/borrowers.Borrower'
        type: array
      history:
        description: Historical records of borrowed books
        items:
          $ref: '#/definitions/borrowers.Borrower'
        type: array
    type: object
  borrowers.UpdateBorrowerRequest:
    properties:
      book_id:
        type: string
      borrow_date:
        type: string
      id:
        type: string
      return_date:
        type: string
      user_id:
        type: string
    type: object
  borrowers.UpdateBorrowerResponse:
    properties:
      borrower:
        $ref: '#/definitions/borrowers.Borrower'
    type: object
  genres.CreateGenreRequest:
    properties:
      name:
        type: string
    type: object
  genres.CreateGenreResponse:
    properties:
      genre:
        $ref: '#/definitions/genres.Genre'
    type: object
  genres.Genre:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  genres.GetGenreByIdResponse:
    properties:
      genre:
        $ref: '#/definitions/genres.Genre'
    type: object
  genres.GetGenresResponse:
    properties:
      genres:
        items:
          $ref: '#/definitions/genres.Genre'
        type: array
    type: object
  genres.UpdateGenreRequest:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  genres.UpdateGenreResponse:
    properties:
      genre:
        $ref: '#/definitions/genres.Genre'
    type: object
info:
  contact: {}
  description: API for managing Library Management System resources
  title: Library Management System API
  version: "1.0"
paths:
...
  /api/borrowers/:
    get:
      description: Retrieve a list of all borrowers
      parameters:
      - in: query
        name: book_id
        type: string
      - in: query
        name: borrow_date
        type: string
      - in: query
        name: return_date
        type: string
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Menu item found
          schema:
            $ref: '#/definitions/borrowers.GetAllBorrowersResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all borrowers
      tags:
      - Borrowers
    post:
      consumes:
      - application/json
      description: Add a new borrower to the library system
      parameters:
      - description: Borrower Data
        in: body
        name: borrower
        required: true
        schema:
          $ref: '#/definitions/borrowers.CreateBorrowerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/borrowers.CreateBorrowerResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new borrower
      tags:
      - Borrowers
  /api/borrowers/{id}:
    delete:
      description: Remove a borrower from the library system by their ID
      parameters:
      - description: Borrower ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Borrower deleted successfully
          schema:
            type: string
        "400":
          description: Borrower ID
          schema:
            type: string
        "404":
          description: Borrower Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete borrower by ID
      tags:
      - Borrowers
    get:
      description: Retrieve details of a borrower by their ID
      parameters:
      - description: Borrower ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/borrowers.GetBorrowerByIdResponse'
        "400":
          description: Borrower ID
          schema:
            type: string
        "404":
          description: Borrower Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get borrower by ID
      tags:
      - Borrowers
    put:
      consumes:
      - application/json
      description: Update details of an existing borrower by their ID
      parameters:
      - description: Borrower ID
        in: path
        name: id
        required: true
        type: string
      - description: Borrower Data
        in: body
        name: borrower
        required: true
        schema:
          $ref: '#/definitions/borrowers.UpdateBorrowerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/borrowers.UpdateBorrowerResponse'
        "400":
          description: Borrower ID
          schema:
            type: string
        "404":
          description: Borrower Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update borrower by ID
      tags:
      - Borrowers
  /api/borrowers/users/{user_id}/borrowed_books:
    get:
      description: Retrieve a list of books borrowed by a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/borrowers.GetAllBorrowersResponse'
        "400":
          description: User ID
          schema:
            type: string
        "404":
          description: Borrower Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get borrowed books by user ID
      tags:
      - Borrowers
  /api/borrowers/users/{user_id}/borrowing_history:
    get:
      description: Retrieve borrowing history of a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/borrowers.GetBorrowingHistoryResponse'
        "400":
          description: Invalid User ID
          schema:
            type: string
        "404":
          description: Borrowing History Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get borrowing history by user ID
      tags:
      - Borrowers
  /api/genres/:
    get:
      description: Retrieve a list of all genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genres.GetGenresResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all genres
      tags:
      - Genres
    post:
      consumes:
      - application/json
      description: Add a new genre to the system
      parameters:
      - description: Genre Data
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/genres.CreateGenreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/genres.CreateGenreResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new genre
      tags:
      - Genres
  /api/genres/{genre_id}/books:
    get:
      description: Retrieve a list of books belonging to a specific genre
      parameters:
      - description: Genre ID
        in: path
        name: genre_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genres.GetGenreByIdResponse'
        "400":
          description: Genre ID
          schema:
            type: string
        "404":
          description: Genre Not Found for Genre
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get books by genre ID
      tags:
      - Genres
  /api/genres/{id}:
    delete:
      description: Remove a genre from the system by its ID
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Borrower deleted successfully
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Genre Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete genre by ID
      tags:
      - Genres
    get:
      description: Retrieve details of a genre by its ID
      parameters:
      - description: Genre Genre ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genres.GetGenreByIdResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Genre Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get genre by ID
      tags:
      - Genres
    put:
      consumes:
      - application/json
      description: Update details of an existing genre by its ID
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: string
      - description: Genre Data
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/genres.UpdateGenreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genres.UpdateGenreResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "404":
          description: Genre Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update genre by ID
      tags:
      - Genres
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"